requirements.txt
requests
ldap3
python-dotenv